Second Commit

+++ app/src/dev/patika/models/PermanentInstructor.java
@@ -8,20 +8,20 @@ import javax.persistence.Table;
 @Table(name = "permanent_instructor")
 public class PermanentInstructor extends Instructor{
     @Column(name = "vr_fixed_salary")
-    private int fixedSalary;
+    private double fixedSalary;
 
-    public PermanentInstructor(String fullName, String address, String phoneNumber, int fixedSalary) {
+    public PermanentInstructor(String fullName, String address, String phoneNumber, double fixedSalary) {
         super(fullName, address, phoneNumber);
         this.fixedSalary = fixedSalary;
     }
 
     public PermanentInstructor() {}
 
-    public int getFixedSalary() {
+    public double getFixedSalary() {
         return fixedSalary;
     }
 
-    public void setFixedSalary(int fixedSalary) {
+    public void setFixedSalary(double fixedSalary) {
         this.fixedSalary = fixedSalary;
     }
 }
